Se terminó la interfaz principal y se arreglaron unos detalles del modelo Se agregaron los campos restantes de id de destinatario y valor en la interfaz y se agrego un botón de cerrar sesión Se agrego el procedimiento almacenado de la base de datos Se acoplo la clase transaccion para que vaya de acuerdo al campo double del valor, en la base de datos

+++ src/modelo/TransaccionDAO.java
@@ -63,7 +63,7 @@ public class TransaccionDAO implements ComunicacionDB <Transaccion> {
 					transaccion.setId(resultados.getInt("id"));
 					transaccion.setIdDestinatario(resultados.getInt("idDestinatario"));
 					transaccion.setIdEmisor(resultados.getInt("idEmisor"));
-					transaccion.setValor(resultados.getInt("valor"));
+					transaccion.setValor(resultados.getDouble("valor"));
 					transacciones.add(transaccion);
 				}
 			} catch (SQLException e) {
